pyproject.toml
[project]
name = "christmas-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "pyyaml>=6.0",
]
